third_party/blink/web_tests/external/wpt/css/css-anchor-position/position-area-self-alignment-overflow.html

<!DOCTYPE html>
<title>CSS Anchor Positioning: position-area positioning - normal alignment overflow</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-1/#alignment">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  /* A 300x300 container with a 100x100 centered anchor */
  #container {
    position: relative;
    width: 300px;
    height: 300px;
  }
  #anchor {
    position: absolute;
    top: 100px;
    left: 100px;
    width: 100px;
    height: 100px;
    anchor-name: --anchor;
  }
  #anchored {
    position: absolute;
    /* The position-areas are 100px*100px. This box will overflow the position area,
       making this element be shifted towards the anchor and overflow the anchor to
       stay within the original containing block. */
    width: 150px;
    height: 150px;
    position-anchor: --anchor;
  }
</style>
<div id="container">
  <div id="anchor"></div>
  <div id="anchored"></div>
</div>
<script>
  function test_position_area(position_area, expected_offsets) {
    anchored.style.positionArea = position_area;

    test(() => {
      assert_equals(anchored.offsetLeft, expected_offsets.left, "Check expected offsetLeft");
      assert_equals(anchored.offsetTop, expected_offsets.top, "Check expected offsetTop");
      assert_equals(anchored.offsetWidth, expected_offsets.width, "Check expected offsetWidth");
      assert_equals(anchored.offsetHeight, expected_offsets.height, "Check expected offsetHeight");
    }, "Offsets for position-area: " + position_area);
  }

  test_position_area("top left", {left:0, top:0, width:150, height:150});
  test_position_area("top center", {left:75, top:0, width:150, height:150});
  test_position_area("top right", {left:150, top:0, width:150, height:150});
  test_position_area("center left", {left:0, top:75, width:150, height:150});
  test_position_area("center right", {left:150, top:75, width:150, height:150});
  test_position_area("bottom left", {left:0, top:150, width:150, height:150});
  test_position_area("bottom center", {left:75, top:150, width:150, height:150});
  test_position_area("bottom right", {left:150, top:150, width:150, height:150});

</script>
